requirements.txt
numpy
ipykernel
wheel
moviepy 
luxai_s2
jax
flax
gym
opencv-python
rlax
juxai-s2